# backend/internal/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Resolve API
  description: >-
    Resolve v0 Core Tickets — a small ticketing system. Mutating endpoints
    accept an optional `X-Actor` header (defaults to `api`) recorded against
    the audit log.
  version: "0.1.0"
servers:
  - url: /
tags:
  - name: tickets
  - name: audit
  - name: stats
  - name: canned-responses
  - name: health

paths:
  /health:
    get:
      tags: [health]
      summary: Service health and uptime
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /tickets:
    post:
      tags: [tickets]
      summary: Create a ticket
      parameters:
        - $ref: "#/components/parameters/XActor"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTicketRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "400":
          $ref: "#/components/responses/BadRequest"
    get:
      tags: [tickets]
      summary: List tickets
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/Status"
        - name: priority
          in: query
          schema:
            $ref: "#/components/schemas/Priority"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticket"

  /tickets/{id}:
    get:
      tags: [tickets]
      summary: Get a ticket by id, including comments
      parameters:
        - $ref: "#/components/parameters/TicketId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "404":
          $ref: "#/components/responses/NotFound"

  /tickets/{id}/status:
    post:
      tags: [tickets]
      summary: Transition a ticket's status
      description: >-
        Only whitelisted transitions are allowed: new→open, open→in_progress,
        in_progress→waiting_customer|resolved, waiting_customer→in_progress,
        resolved→closed. Illegal moves return 400 listing the allowed next
        states.
      parameters:
        - $ref: "#/components/parameters/TicketId"
        - $ref: "#/components/parameters/XActor"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to]
              properties:
                to:
                  $ref: "#/components/schemas/Status"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

  /tickets/{id}/triage:
    post:
      tags: [tickets]
      summary: Record a triage classification result
      description: >-
        Called by the triage service, not end users. medium/high confidence
        results overwrite the ticket's category/priority directly; low
        confidence results are stored as a pending suggestion (see
        pendingCategory/pendingPriority on Ticket) for a human to review via
        POST /tickets/{id}/triage/review instead of being applied.
      parameters:
        - $ref: "#/components/parameters/TicketId"
        - $ref: "#/components/parameters/XActor"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [category, priority, confidence]
              properties:
                category:
                  $ref: "#/components/schemas/Category"
                priority:
                  $ref: "#/components/schemas/Priority"
                confidence:
                  $ref: "#/components/schemas/Confidence"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

  /tickets/{id}/triage/review:
    post:
      tags: [tickets]
      summary: Accept or reject a pending (low-confidence) triage suggestion
      description: >-
        400s if the ticket has no pending suggestion — check
        pendingCategory/pendingPriority on the Ticket before calling.
      parameters:
        - $ref: "#/components/parameters/TicketId"
        - $ref: "#/components/parameters/XActor"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision:
                  type: string
                  enum: [accept, reject]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

  /tickets/{id}/comments:
    post:
      tags: [tickets]
      summary: Add a comment to a ticket
      parameters:
        - $ref: "#/components/parameters/TicketId"
        - $ref: "#/components/parameters/XActor"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

  /tickets/{id}/audit:
    get:
      tags: [tickets, audit]
      summary: List a ticket's audit entries, newest first
      parameters:
        - $ref: "#/components/parameters/TicketId"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

  /audit:
    get:
      tags: [audit]
      summary: List audit entries, optionally filtered by ticket
      parameters:
        - name: ticketId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditEntry"

  /stats:
    get:
      tags: [stats]
      summary: Ticket counts by status/priority and average resolution time
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsResponse"

  /canned-responses:
    post:
      tags: [canned-responses]
      summary: Create a canned response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveCannedResponseRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CannedResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
    get:
      tags: [canned-responses]
      summary: List canned responses, ordered by title ascending
      parameters:
        - $ref: "#/components/parameters/Tag"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CannedResponse"

  /canned-responses/{id}:
    get:
      tags: [canned-responses]
      summary: Get a canned response by id
      parameters:
        - $ref: "#/components/parameters/CannedResponseId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CannedResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [canned-responses]
      summary: Update a canned response
      parameters:
        - $ref: "#/components/parameters/CannedResponseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveCannedResponseRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CannedResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [canned-responses]
      summary: Delete a canned response
      description: >-
        Does not alter any ticket comment previously created by inserting
        this canned response's text — comments store plain text with no
        live reference back to the canned response.
      parameters:
        - $ref: "#/components/parameters/CannedResponseId"
      responses:
        "204":
          description: No Content
        "404":
          $ref: "#/components/responses/NotFound"

components:
  parameters:
    TicketId:
      name: id
      in: path
      required: true
      schema:
        type: string
    XActor:
      name: X-Actor
      in: header
      required: false
      description: Recorded as the audit-log actor. Defaults to `api`.
      schema:
        type: string
        default: api
    Limit:
      name: limit
      in: query
      required: false
      description: Max rows to return. Defaults to 50, clamped to 200.
      schema:
        type: integer
    Offset:
      name: offset
      in: query
      required: false
      description: Rows to skip before collecting the page. Defaults to 0.
      schema:
        type: integer
    CannedResponseId:
      name: id
      in: path
      required: true
      schema:
        type: string
    Tag:
      name: tag
      in: query
      required: false
      description: >-
        Filter to canned responses tagged with this value. Matching is
        case-insensitive; an empty value is treated as no filter.
      schema:
        type: string

  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Ticket not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      example:
        message: "ticket tkt_01 not found"

    HealthResponse:
      type: object
      properties:
        service:
          type: string
          example: resolve
        version:
          type: string
          example: dev
        uptimeSeconds:
          type: number
          format: double

    Priority:
      type: string
      enum: [low, normal, high, urgent]

    Status:
      type: string
      enum: [new, open, in_progress, waiting_customer, resolved, closed]

    Category:
      type: string
      description: Assigned by the triage service; null until triage runs.
      enum: [billing, bug, account_access, feature_request, how_to, other]

    Confidence:
      type: string
      description: >-
        How confident the triage service is in a classification. medium/high
        results apply directly; low results are stored as a pending
        suggestion for a human to review instead.
      enum: [low, medium, high]

    CreateTicketRequest:
      type: object
      required: [subject, description, customerEmail, priority]
      properties:
        subject:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        customerEmail:
          type: string
          format: email
        priority:
          $ref: "#/components/schemas/Priority"

    CreateCommentRequest:
      type: object
      required: [author, body]
      properties:
        author:
          type: string
          minLength: 1
        body:
          type: string
          minLength: 1
        internal:
          type: boolean
          default: false
          description: Internal notes are agent-only, never exposed to customers.

    Comment:
      type: object
      properties:
        id:
          type: string
        author:
          type: string
        body:
          type: string
        internal:
          type: boolean
        at:
          type: string
          format: date-time

    Ticket:
      type: object
      properties:
        id:
          type: string
        subject:
          type: string
        description:
          type: string
        customerEmail:
          type: string
          format: email
        priority:
          $ref: "#/components/schemas/Priority"
        category:
          allOf:
            - $ref: "#/components/schemas/Category"
          nullable: true
        status:
          $ref: "#/components/schemas/Status"
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
          nullable: true
        pendingCategory:
          allOf:
            - $ref: "#/components/schemas/Category"
          nullable: true
          description: A low-confidence triage suggestion awaiting human review.
        pendingPriority:
          allOf:
            - $ref: "#/components/schemas/Priority"
          nullable: true
          description: A low-confidence triage suggestion awaiting human review.

    AuditEntry:
      type: object
      properties:
        actor:
          type: string
        action:
          type: string
          example: ticket.status_changed
        ticketId:
          type: string
        details:
          type: object
          additionalProperties: true
        at:
          type: string
          format: date-time

    AuditPage:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/AuditEntry"
        limit:
          type: integer
        offset:
          type: integer
        hasMore:
          type: boolean

    StatsResponse:
      type: object
      properties:
        total:
          type: integer
        byStatus:
          type: object
          additionalProperties:
            type: integer
        byPriority:
          type: object
          additionalProperties:
            type: integer
        avgResolutionMinutes:
          type: integer
          nullable: true

    SaveCannedResponseRequest:
      type: object
      required: [title, body]
      properties:
        title:
          type: string
          minLength: 1
        body:
          type: string
          minLength: 1
        tags:
          type: array
          items:
            type: string
            minLength: 1
          default: []
          description: >-
            Each tag is stored trimmed and lower-cased; an empty/whitespace
            entry is rejected.

    CannedResponse:
      type: object
      properties:
        id:
          type: string
          example: cr_a1b2c3d4
        title:
          type: string
        body:
          type: string
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time